pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "datasec"
version = "0.2.0"
description = "Personal data protection toolkit — breach scanning, file encryption, OSINT, password audit, hidden volumes, metadata stripping"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "estebanddlc" }]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7",
    "rich>=13.7.0",
    "requests>=2.31.0",
    "cryptography>=42.0.0",
    "schedule>=1.2.0",
    "plyer>=2.1.0",
    "exifread>=3.0.0",
    "Pillow>=10.0.0",
    "pypdf>=4.0.0",
    "mutagen>=1.47.0",
]

[project.scripts]
datasec = "datasec.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["datasec*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
